Refactor BE structure
