+++ api/app/Console/Commands/Stubs/CoreRepositoryInterface.stub
@@ -2,16 +2,6 @@
 
 namespace App\Interfaces\Repository;
 
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Pagination\LengthAwarePaginator;
-
-interface {name}
+interface {name} extends BaseRepositoryInterface
 {
-    public function getModel(): Model;
-    public function list(array $params): Collection | LengthAwarePaginator;
-    public function create(array $data): Model;
-    public function update(array $data, string $uuid): bool;
-    public function delete(string $uuid): bool;
-    public function find(string $uuid): ?Model;
 }
