+++ api/routes/api.php
@@ -1,7 +1,7 @@
 <?php
 
-use App\Http\Controllers\CategoryController;
-use App\Http\Controllers\ManufacturerController;
+use App\Http\Controllers\Category\CategoryController;
+use App\Http\Controllers\Manufacturer\ManufacturerController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
